refactor: Presentation -> Translator로 변경
service 패키지로 이동.
